requirements.txt
babel==2.16.0
colorama==0.4.6
docopt==0.6.2
iniconfig==2.0.0
num2words==0.5.13
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
six==1.16.0